backend/pyproject.toml

[tool.poetry]
name = "wisherr-backend"
version = "0.1.0"
description = "Backend API Wisherr (FastAPI, PostgreSQL, Auth, i18n)"
authors = ["Projet Wisherr"]
package-mode = false


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110"
uvicorn = "^0.29"
sqlmodel = "^0.0.16"
asyncpg = "^0.29"
psycopg2-binary = "^2.9"
python-jose = "^3.3"
passlib = {extras = ["argon2"], version = "^1.7"}
python-multipart = "^0.0.9"
Authlib = "^1.3"
python-dotenv = "^1.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
Pillow = "^10.1.0"
slowapi = "^0.1.9"
pydantic = {extras = ["email"], version = "^2.5"}
alembic = "^1.13"
redis = "^5.0"
websockets = "^12.0"
apscheduler = "^3.10"
prometheus-client = "^0.16.0"

[tool.poetry.dev-dependencies]
pytest = "^8.0"
httpx = "^0.27"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
